Ghetto Synth Repair
This PR adds a few ghetto options for specific steps of synth repair. Most of the steps had ghetto options except for the actual repair surgeries.

With this, it's possible to use a wrench and wirecutters to repair brute and burn respectively, but at a pitiful 35% base success rate and a near-doubling of the time taken to perform the action.