Bearcat away map rework. (Two different map formats for the same away site) Edition.
This adds a functional Overmap Shuttle to the Bearcat away map. For some bizarre reason the two map files our of the opposite mapping format. I Map merged and reconverted both of the map files to how they were. I also added in mining equipment and suit cyclers to the Bearcat ship to allow for continuations  power generation without having to rely on the Dagon for help.