Changes the cost of stims and zerk
Stimpack: 3TC > 10TC
Oxycodone is very powerful, and 10 units of it mixed with 5 of inaprovaline is enough to keep someone standing long after they should have collapsed. Considering the autoinjector for some reason contains three times the volume of normal autoinjectors, 10TC is a much fairer price for what is basically three to four minutes of invulnerability inside of a hardsuit. Traitors shouldn't be able to carry eight of them in a wallet and walk around with their guts hanging out, perfectly happy.

Zerk: 5TC > 15TC
For reasons similar to my reasoning with stims, three doses of a synaptizine and hyperzine mix shouldn't be as cheap as it is, especially when there's nothing preventing you from pairing it with stims and just becoming immortal. More than one combo of that should require more TC commitment than the loose change you had after buying a gun, ammo and armor.