The Bulldog can load non-frangible mags again.
It's currently a worse SMG since the frangibles do nothing the moment the tiniest bit of armor pokes them. Also, the security Bulldog is a way worse variant of the genuine bullpup rifle, it has: jam chance, no burst fire, needs commandauthorization/red alert to fire & it has overall worse stats when compared to the bullpup. This should, finally make the weapon something that isn't taken from the SEC-ARM jsut because it looks cool.