Buffs the CoSs Hazard RIG to actually be useful, and used.
Projectile protection from, ARMOR_BALLISTIC_PISTOL to ARMOR_BALLISTIC_RIFLE
Laser protection from, ARMOR_LASER_HANDGUNS to ARMOR_LASER_MAJOR.

Feedback may or may not be required. It's only a slight buff. However this will allow the CoS to be somewhat capable in the Hardsuit, instead of opting for actual armor which is better off.