Removes all foundation pistol spawns.
Apparently, we had two of these. Now we have none. Psionic powers rely on a fairly high amount of RNG to acquire, and then you need to make more RNG rolls to determine if you get a decent school at a decent level. Far too many RNG checks for powers that can, for the most part, be completely replaced with a few tools or items.