Light Rig added to MechFab
Honestly, this is me mostly testing code, and getting the grips with the spaghetti code that is SS13 development. This simply adds the Light Rig to the MechFab. The light rig is not EVA proof, and provides very low protection alike to the Null suit that the MechFab also provides. Its only slightly better then the Null suit, therefore has a research requirement, and slightly higher material cost. Again throw ya feedback at me, if this is shit then please say so. I am simply doing this to learn.