Adds armor plates to loadout
Simply adds light and medium armor plates to the loadout for faction and job specific roles that require them. We can already get light and medium armor attachments such as arm/leg pads, and arm/leg guards but no plates for the plate carrier which to be me doesn't make sense. Since we're a military vessel a plate carrier with a actual plate is near standard issue for most aboard the ship.