Removing CorpoSec
God help me, I am prepared for to be ass fisted for this. This PR removes CorpoSec from the Security. Considering our current lore regarding the Corpo War, and that fact that the Dagon is now a military vessel, it does not make sense that we would have Corpos in one of the most important jobs on the ship. Especially considering why would Sol Command hire expensive Corpo Guards when they have a supply of loyal, cheap military personnel to fill such roles.