Fucking Irkalla starts 'coding' or rather reflavoring shit again somebody stop them before they stealth buff cargo for joe jones gimmicks
About The Pull Request
This reflavors the Tiro equipment, part of a series to make them more into COMBINE or ADVENT troopers instead of what it was before. Will change later (not today) the description for the particle magnum as well and alter the "par" sigil to be something else.

Par means "equal". Somehow I was not aware of that. That is on par of doing.. wait I get it.

Auctus means  "augmented"/enhanced. The suit already got a few nerfs in so its more on par with Alate suits. Could still be changed by somebody that knows balancing better than I do to make it weaker. The suit is basicly now an "augmented" Tiro suit. Mentions of being the Gynes personal bodyguard are squashed. They are now more akin to combine elite or advent officers. Which fits them better if you ask me but dunno. Just thought this would suit it better. 

Why It's Good For The Game
Swear to god I am going to beat Tiros with a crowbar clustertool if that isnt going to be reflavored into something more akin to COMBINE or ADVENT. That and the fucking alate that always puts ~ at the end of every sentence. You know who you are.

Did You Test It?
No, I just changed description names.

Authorship
Irkalla Epsilon

Changelog
🆑
tweak: Combinefies/Adherentifies Ascent Tiros and makes them less ERPy with their gears flavortext. This is a multi step series. Ascent Par suit to Auctus. Pars are now "Auctus" or "Augmented". No more bodyguard flavortext on this particular suit. To do: Change particle magnum and par emblem to something similar. Fuck coding. I go back make materials and shit.
/🆑