Initial Lace Rebalance - - - IMPORTANT - - - Laces, now rather than killing their user immediately, provide an avenue of RP, keep the player in the round and are no longer an incredible detriment. Mostly. Destruction of the lace will remove all skills, temporarily stun you via seizure, give mild brain damage and force hallucinations for a hot minute. To compensate, it has been made far harder to destroy and still retains the 0.5 damage modifier.
To understand why, it's assumed that the user's conscious, stored normally in the lace to some degree, would sustain tremendous damage upon destruction of said device. You don't so much lose yourself as it is your brain being forcefully rewired, the implant being housed between the cerebellum and medulla oblongata and in a prime area for such.
Alternatively, refer to the in game description: "By way of violent neuron rearrangement, you're practically a husk."

This will likely need tweaks, but is absolutely more in line with what we're aiming for.
- - -