I made weed better
## About The Pull Request
Increases yield on cannabis to 5. It already has a long maturation and production so I made it's yield slightly lower than ambrosia so you don't need to dump in an entire bottle of robust harvest for it not to suck
Makes them show up on examine.
## Why It's Good For The Game
Makes producing cannabis, which is only really an RP flavor plant less of a chore
## Did You Test It?
Ye
## Authorship
YuiY1997 CKEY -YuiY1997
## Changelog

:cl:
tweak:changed yield of cannabis from 1 to 5
/:cl: