Buffs shields.
Shields are a funny meme and nobody uses them outside of the rule of cool. When they were first introduced you could accurately fire longarms while using them. This is no longer the case, so buffing the block chance, amount, and slightly reducing the slowdown seemed like a good idea.